requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
mysql-connector-python
python-jose
passlib
python-dotenv
pydantic-settings
bcrypt
python-multipart
